# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PaymentType = typeof PaymentType[keyof typeof PaymentType];


export const PaymentType = {
  cash: 'cash',
  card: 'card',
  transfer: 'transfer',
} as const;

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  income: 'income',
  expense: 'expense',
} as const;

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  operator: 'operator',
} as const;

export interface User {
  id: number;
  clerkUserId: string;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface UserRoleUpdate {
  role: UserRole;
}

export interface Client {
  id: number;
  name: string;
  territory: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  responsiblePerson: string | null;
  /** Positive = haqdorlik (they owe us), negative = qarzdorlik (we owe them) */
  balance: number;
  createdAt: string;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  territory: string;
  phone?: string;
  responsiblePerson?: string;
}

export interface ClientUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  territory?: string;
  phone?: string;
  responsiblePerson?: string;
}

export interface Transaction {
  id: number;
  clientId: number;
  clientName: string;
  type: TransactionType;
  /** @minimum 0 */
  amount: number;
  date: string;
  paymentType: PaymentType;
  responsiblePerson: string;
  /** @nullable */
  comment: string | null;
  createdAt: string;
}

export interface TransactionInput {
  clientId: number;
  type: TransactionType;
  /** @exclusiveMinimum 0 */
  amount: number;
  date: string;
  paymentType: PaymentType;
  /** @minLength 1 */
  responsiblePerson: string;
  comment?: string;
}

export interface TransactionUpdate {
  clientId?: number;
  type?: TransactionType;
  /** @exclusiveMinimum 0 */
  amount?: number;
  date?: string;
  paymentType?: PaymentType;
  /** @minLength 1 */
  responsiblePerson?: string;
  comment?: string;
}

export interface CashFlowPoint {
  date: string;
  income: number;
  expense: number;
}

export interface DashboardSummary {
  /** Jami haqdorlik */
  totalReceivable: number;
  /** Jami qarzdorlik */
  totalPayable: number;
  netBalance: number;
  cashFlow: CashFlowPoint[];
}

export interface AktSverkaRow {
  date: string;
  /** @nullable */
  comment: string | null;
  responsiblePerson: string;
  paymentType: PaymentType;
  /** Kirim (mijoz to'ladi) */
  debit: number;
  /** Chiqim (mijozga berildi) */
  credit: number;
  balance: number;
}

export interface AktSverkaResult {
  clientId: number;
  clientName: string;
  periodStart: string;
  periodEnd: string;
  openingBalance: number;
  closingBalance: number;
  rows: AktSverkaRow[];
}

export type ListClientsParams = {
search?: string;
territory?: string;
};

export type ListTransactionsParams = {
clientId?: number;
search?: string;
paymentType?: PaymentType;
responsiblePerson?: string;
dateFrom?: string;
dateTo?: string;
};

export type GetAktSverkaParams = {
clientId: number;
from: string;
to: string;
};

export type ExportClientsParams = {
search?: string;
territory?: string;
};

export type ExportTransactionsParams = {
clientId?: number;
search?: string;
paymentType?: PaymentType;
responsiblePerson?: string;
dateFrom?: string;
dateTo?: string;
};

export type ExportAktSverkaExcelParams = {
clientId: number;
from: string;
to: string;
};

export type ExportAktSverkaPdfParams = {
clientId: number;
from: string;
to: string;
};