# lib/api-zod/src/generated/types/paymentType.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PaymentType = typeof PaymentType[keyof typeof PaymentType];


export const PaymentType = {
  cash: 'cash',
  card: 'card',
  transfer: 'transfer',
} as const;